+++ src/angular-cesium-entities-editor/components/polylines-editor/polylines-editor.component.ts
@@ -144,6 +144,21 @@ export class PolylinesEditorComponent implements OnDestroy {
 				}
 				break;
 			}
+			case EditActions.DRAG_SHAPE: {
+				const polyline = this.polylinesManager.get(update.id);
+				if (polyline && polyline.enableEdit) {
+					polyline.moveShape(update.draggedPosition, update.updatedPosition)
+				}
+				break;
+			}
+			
+			case EditActions.DRAG_SHAPE_FINISH: {
+				const polyline = this.polylinesManager.get(update.id);
+				if (polyline && polyline.enableEdit) {
+					polyline.endMoveShape();
+				}
+				break;
+			}
 			default: {
 				return;
 			}